split dashes into separate files, add car 6 dash
